+++ loco-new/base_template/src/mailers/auth/magic_link/text.t
@@ -1,0 +1,2 @@
+Magic link with this link: 
+{{host}}/api/auth/magic-link/{{token}}